moved config_manager to it's own directory
