+++ powerberry-dsp/Makefile
@@ -1,5 +1,5 @@
 
-Files=./src/*.cpp ./src/spi/*.cpp ./src/adc/*.cpp ./src/filters/*.cpp ./src/controller/*.cpp ./src/datastorage/*.cpp
+Files=./src/*.cpp ./src/spi/*.cpp ./src/adc/*.cpp ./src/filters/*.cpp ./src/controller/*.cpp ./src/datastorage/*.cpp ./src/config_manager/*.cpp
 Libraries=/usr/local/lib/libredis++.a /usr/local/lib/libhiredis.a -l bcm2835
 Includes= ./src
 #.PHONY : ./bin/powerberry-dsp
